split up insert script into module, create scripts

+++ test/test_insert.py
@@ -1,36 +1,49 @@
 import json
 import sys
 from datetime import datetime
+from pathlib import Path
 
 import pytest
 from sqlalchemy import create_engine, func, select
 from sqlalchemy.orm import Session
 
 sys.path.append("S:/Dokumente/Code/ndv-elo/src")
+# Linux parent directory
+sys.path.append(str(Path(".").absolute()))
 import os
 
-import schema
-from insert_data import (
+from src.insert import (
     populate_clubs_and_teams,
     populate_competitions,
     populate_players,
     populate_teammatches,
 )
+from src.schema import (
+    Base,
+    Club,
+    Competition,
+    DoublesMatch,
+    Player,
+    SinglesMatch,
+    TeamMatch,
+)
 
 DB_PATH = "test.db"
 DB_URL = f"sqlite:///{DB_PATH}"
 TEST_DATA_PATH = "./test/testdata.json"
+competition = "Kreisliga 5"
+association = "DBH"
 
 with open(TEST_DATA_PATH, "r") as f:
-    TEST_DATA = json.load(f)
-    TEST_DATA["season"] = datetime.fromisoformat(TEST_DATA["season"])
+    data = json.load(f)
+    data["season"] = datetime.fromisoformat(data["season"])
+
 
 
 @pytest.fixture(scope="session")
 def db_engine():
-    os.remove(DB_PATH)
-    from schema import Base
-
+    if os.path.exists(DB_PATH):
+        os.remove(DB_PATH)
     engine = create_engine(DB_URL)
     Base.metadata.create_all(engine)
     return engine
@@ -38,104 +51,100 @@ def db_engine():
 
 def test_duplicate_insert_clubs(db_engine):
     with Session(db_engine) as session:
-        stmt = select(func.count()).select_from(schema.Club)
+        stmt = select(func.count()).select_from(Club)
         result = session.execute(stmt)
         n_clubs_0 = result.scalar()
     # trunk-ignore(bandit/B101)
     assert n_clubs_0 == 0
     populate_clubs_and_teams(
-        db_engine, TEST_DATA["DBH"]["Kreisliga 5"]["clubs_teams"], TEST_DATA["season"]
+        db_engine, data[association][competition]["clubs_teams"], data["season"]
     )
     with Session(db_engine) as session:
-        stmt = select(func.count()).select_from(schema.Club)
+        stmt = select(func.count()).select_from(Club)
         result = session.execute(stmt)
         n_clubs_1 = result.scalar()
     # trunk-ignore(bandit/B101)
-    assert n_clubs_1 == len(TEST_DATA["DBH"]["Kreisliga 5"]["clubs_teams"])
+    assert n_clubs_1 == len(data[association][competition]["clubs_teams"])
     populate_clubs_and_teams(
-        db_engine, TEST_DATA["DBH"]["Kreisliga 5"]["clubs_teams"], TEST_DATA["season"]
+        db_engine, data[association][competition]["clubs_teams"], data["season"]
     )
     with Session(db_engine) as session:
-        stmt = select(func.count()).select_from(schema.Club)
+        stmt = select(func.count()).select_from(Club)
         result = session.execute(stmt)
         n_clubs_2 = result.scalar()
     # trunk-ignore(bandit/B101)
     assert n_clubs_1 == n_clubs_2
 
 
-
-
 def test_duplicate_insert_competitions(db_engine):
     with Session(db_engine) as session:
-        stmt = select(func.count()).select_from(schema.Competition)
+        stmt = select(func.count()).select_from(Competition)
         result = session.execute(stmt)
         n_comps_0 = result.scalar()
     # trunk-ignore(bandit/B101)
     assert n_comps_0 == 0
-    populate_competitions(
-        db_engine, TEST_DATA["crawled_competitions"], TEST_DATA["season"]
-    )
+    populate_competitions(db_engine, data["crawled_competitions"], data["season"])
     with Session(db_engine) as session:
-        stmt = select(func.count()).select_from(schema.Competition)
+        stmt = select(func.count()).select_from(Competition)
         result = session.execute(stmt)
         n_comps_1 = result.scalar()
     # trunk-ignore(bandit/B101)
-    assert n_comps_1 == len(TEST_DATA["crawled_competitions"])
-    populate_competitions(
-        db_engine, TEST_DATA["crawled_competitions"], TEST_DATA["season"]
-    )
+    assert n_comps_1 == len(data["crawled_competitions"])
+    populate_competitions(db_engine, data["crawled_competitions"], data["season"])
     with Session(db_engine) as session:
-        stmt = select(func.count()).select_from(schema.Competition)
+        stmt = select(func.count()).select_from(Competition)
         result = session.execute(stmt)
         n_comps_2 = result.scalar()
     # trunk-ignore(bandit/B101)
     assert n_comps_1 == n_comps_2
 
+
 def test_duplicate_insert_players(db_engine):
     with Session(db_engine) as session:
-        stmt = select(func.count()).select_from(schema.Player)
+        stmt = select(func.count()).select_from(Player)
         result = session.execute(stmt)
         n_players_0 = result.scalar()
     # trunk-ignore(bandit/B101)
     assert n_players_0 == 0
     populate_players(
         db_engine,
-        TEST_DATA["DBH"]["Kreisliga 5"]["players"],
-        association="DBH",
-        competition="Kreisliga 5",
-        season=TEST_DATA["season"],
+        data[association][competition]["players"],
+        association=association,
+        competition=competition,
+        season=data["season"],
     )
     with Session(db_engine) as session:
-        stmt = select(func.count()).select_from(schema.Player)
+        stmt = select(func.count()).select_from(Player)
         result = session.execute(stmt)
         n_players_1 = result.scalar()
     # trunk-ignore(bandit/B101)
-    assert n_players_1 == len(TEST_DATA["DBH"]["Kreisliga 5"]["players"])
+    assert n_players_1 == len(data[association][competition]["players"])
     populate_players(
         db_engine,
-        TEST_DATA["DBH"]["Kreisliga 5"]["players"],
-        association="DBH",
-        competition="Kreisliga 5",
-        season=TEST_DATA["season"],
+        data[association][competition]["players"],
+        association=association,
+        competition=competition,
+        season=data["season"],
     )
     with Session(db_engine) as session:
-        stmt = select(func.count()).select_from(schema.Player)
+        stmt = select(func.count()).select_from(Player)
         result = session.execute(stmt)
         n_players_2 = result.scalar()
     # trunk-ignore(bandit/B101)
     assert n_players_1 == n_players_2
 
+
 def test_duplicate_insert_teammatches(db_engine):
     with Session(db_engine) as session:
-        stmt_double = select(func.count()).select_from(schema.DoublesMatch)
+        stmt_double = select(func.count()).select_from(DoublesMatch)
         result_double = session.execute(stmt_double)
         n_double_0 = result_double.scalar()
 
-        stmt_single = select(func.count()).select_from(schema.SinglesMatch)
+        stmt_single = select(func.count()).select_from(SinglesMatch)
         result_single = session.execute(stmt_single)
         n_single_0 = result_single.scalar()
 
-        stmt_teammatch = select(func.count()).select_from(schema.TeamMatch)
+        stmt_teammatch = select(func.count()).select_from(TeamMatch)
         result_teammatch = session.execute(stmt_teammatch)
         n_teammatch_0 = result_teammatch.scalar()
     # trunk-ignore(bandit/B101)
@@ -146,51 +155,51 @@ def test_duplicate_insert_teammatches(db_engine):
     assert n_teammatch_0 == 0
     populate_teammatches(
         db_engine,
-        TEST_DATA["DBH"]["Kreisliga 5"]["team_matches"],
-        TEST_DATA["DBH"]["Kreisliga 5"]["matches"],
-        season=TEST_DATA["season"],
+        data[association][competition]["team_matches"],
+        data[association][competition]["matches"],
+        season=data["season"],
     )
     with Session(db_engine) as session:
-        stmt_double = select(func.count()).select_from(schema.DoublesMatch)
+        stmt_double = select(func.count()).select_from(DoublesMatch)
         result_double = session.execute(stmt_double)
         n_double_1 = result_double.scalar()
 
-        stmt_single = select(func.count()).select_from(schema.SinglesMatch)
+        stmt_single = select(func.count()).select_from(SinglesMatch)
         result_single = session.execute(stmt_single)
         n_single_1 = result_single.scalar()
 
-        stmt_teammatch = select(func.count()).select_from(schema.TeamMatch)
+        stmt_teammatch = select(func.count()).select_from(TeamMatch)
         result_teammatch = session.execute(stmt_teammatch)
         n_teammatch_1 = result_teammatch.scalar()
     # trunk-ignore(bandit/B101)
-    assert n_teammatch_1 == len(TEST_DATA["DBH"]["Kreisliga 5"]["team_matches"])
+    assert n_teammatch_1 == len(data[association][competition]["team_matches"])
 
-    n_matches = sum([len(m) for m in TEST_DATA["DBH"]["Kreisliga 5"]["matches"]])
+    n_matches = sum([len(m) for m in data[association][competition]["matches"]])
     # trunk-ignore(bandit/B101)
     assert (
         n_double_1 + n_single_1 == n_matches - 1
     )  # corrupted data has one duplicate match
     populate_teammatches(
         db_engine,
-        TEST_DATA["DBH"]["Kreisliga 5"]["team_matches"],
-        TEST_DATA["DBH"]["Kreisliga 5"]["matches"],
-        season=TEST_DATA["season"],
+        data[association][competition]["team_matches"],
+        data[association][competition]["matches"],
+        season=data["season"],
     )
     with Session(db_engine) as session:
-        stmt_double = select(func.count()).select_from(schema.DoublesMatch)
+        stmt_double = select(func.count()).select_from(DoublesMatch)
         result_double = session.execute(stmt_double)
         n_double_2 = result_double.scalar()
 
-        stmt_single = select(func.count()).select_from(schema.SinglesMatch)
+        stmt_single = select(func.count()).select_from(SinglesMatch)
         result_single = session.execute(stmt_single)
         n_single_2 = result_single.scalar()
 
-        stmt_teammatch = select(func.count()).select_from(schema.TeamMatch)
+        stmt_teammatch = select(func.count()).select_from(TeamMatch)
         result_teammatch = session.execute(stmt_teammatch)
         n_teammatch_2 = result_teammatch.scalar()
     # trunk-ignore(bandit/B101)
     assert n_teammatch_2 == n_teammatch_1
 
-    n_matches = sum([len(m) for m in TEST_DATA["DBH"]["Kreisliga 5"]["matches"]])
+    n_matches = sum([len(m) for m in data[association][competition]["matches"]])
     # trunk-ignore(bandit/B101)
     assert n_double_1 + n_single_1 == n_double_2 + n_single_2
